Angular material

+++ src/main/java/com/elderbyte/vidada/service/MediaLibraryService.java
@@ -22,6 +22,7 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
@@ -45,7 +46,6 @@ public class MediaLibraryService  {
     private final TagService tagService;
 	private final MediaLibraryRepository repository;
     private final MediaRepository mediaRepository;
-    private final MediaSourceRepository mediaSourceRepository;
 
 	private final EventHandlerEx<EventArgsG<MediaLibrary>> libraryAddedEvent = new EventHandlerEx<>();
     private final EventHandlerEx<EventArgsG<MediaLibrary>> libraryRemovedEvent = new EventHandlerEx<>();
@@ -75,13 +75,11 @@ public class MediaLibraryService  {
 	public MediaLibraryService(
         MediaLibraryRepository repository,
         MediaRepository mediaRepository,
-        MediaSourceRepository mediaSourceRepository,
         TagService tagService) {
 
         this.tagService = tagService;
 		this.repository = repository;
         this.mediaRepository = mediaRepository;
-        this.mediaSourceRepository = mediaSourceRepository;
 	}
 
     /***************************************************************************
@@ -111,7 +109,7 @@ public class MediaLibraryService  {
 
     @Transactional
 	public void addLibrary(final MediaLibrary lib){
-        repository.save(lib);
+        save(lib);
         onLibraryAdded(lib);
         libraryAddedEvent.fireEvent(this, EventArgsG.build(lib));
 	}
@@ -120,50 +118,40 @@ public class MediaLibraryService  {
 	public void removeLibrary(final MediaLibrary library) {
         MediaLibrary libToDelete = repository.findOne(library.getId());
 
+        if(libToDelete != null){
+            // Find all medias which have this library
+            List<MediaItem> mediaItems = mediaRepository.queryByLibrary(library);
 
-        // TODO Remove all Sources which used this library
+            for (MediaItem mediaItem : mediaItems) {
+                for (MediaSource source : new HashSet<>(mediaItem.getSources())) {
+                    if(source.getParentLibrary().equals(library)) {
+                        mediaItem.getSources().remove(source);
+                    }
+                }
 
-        // Find all medias which have this library
-        List<MediaSource> sourcesToDelete = new ArrayList<>();
-        List<MediaItem> mediaItems = mediaRepository.queryByLibrary(library);
-
-        for (MediaItem mediaItem : mediaItems) {
-            Set<MediaSource> sources = mediaItem.getSources();
-            for (MediaSource source : sources) {
-                if(source.getParentLibrary().equals(library)) {
-                    mediaItem.getSources().remove(source);
+                // Check if the media has any sources left
+                if(mediaItem.getSources().isEmpty()){
+                    // The media has no sources left
+                    // TODO We can delete it now, but then we loose all general info - is this intended?
+                    mediaRepository.delete(mediaItem);
                 }
             }
 
-            // Check if the media has any sources left
-            if(mediaItem.getSources().isEmpty()){
-                // The media has no sources left
-                // TODO We can delete it now, but then we loose all general info - is this intended?
-                mediaRepository.delete(mediaItem);
-            }
+            repository.delete(libToDelete);
+            libraryRemovedEvent.fireEvent(this, EventArgsG.build(library));
+        }else{
+            log.warn("Could not delete library " + library + " since it was not found in the database!");
         }
-
-        // Now delete all media sources...
-        mediaSourceRepository.deleteInBatch(sourcesToDelete);
-
-        repository.delete(libToDelete);
-        libraryRemovedEvent.fireEvent(this, EventArgsG.build(library));
 	}
 
     @Transactional
-	public void update(final MediaLibrary lib) {
+	public void save(final MediaLibrary lib) {
+        log.info("Saving media library " + lib);
         repository.save(lib);
 	}
 
     @Transactional
-	public MediaLibrary findLibrary(final ResourceLocation file) {
-
-        throw new NotImplementedException();
-        //return repository.queryByLocation(file); // TODO implement this in repository...
-	}
-
-    @Transactional
-	public MediaLibrary getById(final int id) {
+	public MediaLibrary findById(final int id) {
 		return repository.findOne(id);
 	}
